Validate index files
