+++ package.json
@@ -5,7 +5,8 @@
     "scripts": {
         "generate-schemas": "tsx scripts/generate-schemas.ts",
         "copyright": "tsx scripts/copyright.ts",
-        "generate-index-json": "tsx scripts/generate-index-json.ts"
+        "generate-index-json": "tsx scripts/generate-index-json.ts",
+        "validate-index": "tsx scripts/validate-index.ts"
     },
     "repository": {
         "type": "git",
@@ -22,6 +23,7 @@
     "dependencies": {
         "@heroicons/react": "^2.0.18",
         "@octokit/rest": "^20.0.1",
+        "ajv": "^8.12.0",
         "ansi-colors": "^4.1.3",
         "classnames": "^2.3.2",
         "date-fns": "^2.30.0",
